+++ NEWS.md
@@ -8,6 +8,7 @@
 ## Bug Fixes
 
 * Fixed a bug that occurred when calling `predict(object, newdata = newdata, block = TRUE, ...)` if at least one level of a random effect or partition factor from `data` (used to fit `object`) was not present in `newdata`.
+* Fixed a bug that occurred when calling `loocv(object, local = TRUE, ...)` if at least one level of a random effect or partition factor from `data` (used to fit `object`) was represented by only one observation in `data`. 
 
 # spmodel 0.12.0